+++ libraries/common/src/main/java/androidx/media3/common/MediaMetadata.java
@@ -91,6 +91,7 @@ public final class MediaMetadata {
     @Nullable private CharSequence compilation;
     @Nullable private CharSequence station;
     @Nullable private @MediaType Integer mediaType;
+    @Nullable private String playlistId;
     @Nullable private Bundle extras;
     private ImmutableList<String> supportedCommands;
 
@@ -136,6 +137,7 @@ public final class MediaMetadata {
       this.compilation = mediaMetadata.compilation;
       this.station = mediaMetadata.station;
       this.mediaType = mediaMetadata.mediaType;
+      this.playlistId = mediaMetadata.playlistId;
       this.supportedCommands = mediaMetadata.supportedCommands;
       this.extras = mediaMetadata.extras;
     }
@@ -464,6 +466,19 @@ public final class MediaMetadata {
       return this;
     }
 
+    /**
+     * Sets the playlist ID.
+     *
+     * @throws IllegalArgumentException if {@code playlistId} is empty.
+     */
+    @CanIgnoreReturnValue
+    @UnstableApi
+    public Builder setPlaylistId(@Nullable String playlistId) {
+      checkArgument(playlistId == null || !playlistId.isEmpty());
+      this.playlistId = playlistId;
+      return this;
+    }
+
     /** Sets the extras {@link Bundle}. */
     @CanIgnoreReturnValue
     public Builder setExtras(@Nullable Bundle extras) {
@@ -637,6 +652,9 @@ public final class MediaMetadata {
       }
       if (mediaMetadata.mediaType != null) {
         setMediaType(mediaMetadata.mediaType);
+      }
+      if (mediaMetadata.playlistId != null) {
+        setPlaylistId(mediaMetadata.playlistId);
       }
       if (mediaMetadata.extras != null) {
         setExtras(mediaMetadata.extras);
@@ -1173,6 +1191,9 @@ public final class MediaMetadata {
   /** Optional {@link MediaType}. */
   @Nullable public final @MediaType Integer mediaType;
 
+  /** Optional playlist ID. */
+  @UnstableApi @Nullable public final String playlistId;
+
   /**
    * Optional extras {@link Bundle}.
    *
@@ -1243,6 +1264,7 @@ public final class MediaMetadata {
     this.compilation = builder.compilation;
     this.station = builder.station;
     this.mediaType = mediaType;
+    this.playlistId = builder.playlistId;
     this.supportedCommands = builder.supportedCommands;
     this.extras = builder.extras;
   }
@@ -1297,6 +1319,7 @@ public final class MediaMetadata {
         && TextUtils.equals(compilation, that.compilation)
         && TextUtils.equals(station, that.station)
         && Objects.equals(mediaType, that.mediaType)
+        && Objects.equals(playlistId, that.playlistId)
         && Objects.equals(supportedCommands, that.supportedCommands)
         && ((extras == null) == (that.extras == null));
   }
@@ -1340,7 +1363,8 @@ public final class MediaMetadata {
         station,
         mediaType,
         extras == null,
-        supportedCommands);
+        supportedCommands,
+        playlistId);
   }
 
   private static final String FIELD_TITLE = Util.intToStringMaxRadix(0);
@@ -1379,6 +1403,7 @@ public final class MediaMetadata {
   private static final String FIELD_DURATION_MS = Util.intToStringMaxRadix(33);
   private static final String FIELD_SUPPORTED_COMMANDS = Util.intToStringMaxRadix(34);
   private static final String FIELD_DISC_SUBTITLE = Util.intToStringMaxRadix(35);
+  private static final String FIELD_PLAYLIST_ID = Util.intToStringMaxRadix(36);
   private static final String FIELD_EXTRAS = Util.intToStringMaxRadix(1000);
 
   // Use a fairly lenient threshold for sending byte array to legacy processes that don't support
@@ -1516,6 +1541,9 @@ public final class MediaMetadata {
     if (mediaType != null) {
       bundle.putInt(FIELD_MEDIA_TYPE, mediaType);
     }
+    if (playlistId != null) {
+      bundle.putString(FIELD_PLAYLIST_ID, playlistId);
+    }
     if (!supportedCommands.isEmpty()) {
       bundle.putStringArrayList(FIELD_SUPPORTED_COMMANDS, new ArrayList<>(supportedCommands));
     }
@@ -1561,6 +1589,7 @@ public final class MediaMetadata {
         .setGenre(bundle.getCharSequence(FIELD_GENRE))
         .setCompilation(bundle.getCharSequence(FIELD_COMPILATION))
         .setStation(bundle.getCharSequence(FIELD_STATION))
+        .setPlaylistId(bundle.getString(FIELD_PLAYLIST_ID))
         .setExtras(convertToNullIfInvalid(bundle.getBundle(FIELD_EXTRAS)));
 
     if (bundle.containsKey(FIELD_ARTWORK_DATA)) {