+++ libraries/session/src/main/java/androidx/media3/session/MediaConstants.java
@@ -481,6 +481,15 @@ public final class MediaConstants {
       "androidx.media3.session.EXTRAS_KEY_MEDIA_TYPE_COMPAT";
 
   /**
+   * {@link Bundle} key used to indicate the {@link MediaMetadata#playlistId} in the legacy {@code
+   * android.support.v4.media.MediaDescriptionCompat} as a String extra and as a String value in
+   * {@code android.support.v4.media.MediaMetadataCompat}.
+   */
+  @UnstableApi
+  public static final String EXTRAS_KEY_PLAYLIST_ID =
+      "androidx.media3.session.EXTRAS_KEY_PLAYLIST_ID";
+
+  /**
    * {@link Bundle} key used to indicate the {@link CommandButton.Icon} in the extras of the
    * platform {@link android.media.session.PlaybackState.CustomAction}. The corresponding value
    * should be one of the {@code CommandButton.ICON_} integer constants.